+++ examples/all-paging/src/tag/dto/tag.dto.ts
@@ -1,0 +1,24 @@
+import { GraphQLISODateTime, ID, ObjectType } from '@nestjs/graphql'
+import { FilterableField, OffsetConnection, QueryOptions } from '@ptc-org/nestjs-query-graphql'
+
+import { TodoItemDTO } from '../../todo-item/dto/todo-item.dto'
+
+@ObjectType('Tag')
+@QueryOptions({ enableTotalCount: true })
+@OffsetConnection('todoItems', () => TodoItemDTO, {
+  update: { enabled: true },
+  remove: { enabled: true }
+})
+export class TagDTO {
+  @FilterableField(() => ID)
+  id!: number
+
+  @FilterableField()
+  name!: string
+
+  @FilterableField(() => GraphQLISODateTime)
+  created!: Date
+
+  @FilterableField(() => GraphQLISODateTime)
+  updated!: Date
+}