+++ packages/core/src/services/assembler-query.service.ts
@@ -28,26 +28,26 @@ export class AssemblerQueryService<DTO, Entity, C = DeepPartial<DTO>, CE = DeepP
     readonly queryService: QueryService<Entity, CE, UE>
   ) {}
 
-  public addRelations<Relation>(
+  public async addRelations<Relation>(
     relationName: string,
     id: string | number,
     relationIds: (string | number)[],
     opts?: ModifyRelationOptions<DTO, Relation>
   ): Promise<DTO> {
-    return this.assembler.convertAsyncToDTO(
-      this.queryService.addRelations(relationName, id, relationIds, this.convertModifyRelationsOptions(opts))
+    return this.assembler.convertToDTO(
+      await this.queryService.addRelations(relationName, id, relationIds, this.convertModifyRelationsOptions(opts))
     )
   }
 
-  public createMany(items: C[]): Promise<DTO[]> {
+  public async createMany(items: C[]): Promise<DTO[]> {
     const { assembler } = this
-    const converted = assembler.convertToCreateEntities(items)
-    return this.assembler.convertAsyncToDTOs(this.queryService.createMany(converted))
+    const converted = await assembler.convertToCreateEntities(items)
+    return this.assembler.convertToDTOs(await this.queryService.createMany(converted))
   }
 
-  public createOne(item: C): Promise<DTO> {
-    const c = this.assembler.convertToCreateEntity(item)
-    return this.assembler.convertAsyncToDTO(this.queryService.createOne(c))
+  public async createOne(item: C): Promise<DTO> {
+    const c = await this.assembler.convertToCreateEntity(item)
+    return this.assembler.convertToDTO(await this.queryService.createOne(c))
   }
 
   public async deleteMany(filter: Filter<DTO>): Promise<DeleteManyResponse> {
@@ -55,8 +55,8 @@ export class AssemblerQueryService<DTO, Entity, C = DeepPartial<DTO>, CE = DeepP
     return this.queryService.deleteMany(this.assembler.convertQuery({ filter }).filter)
   }
 
-  public deleteOne(id: number | string, opts?: DeleteOneOptions<DTO>): Promise<DTO> {
-    return this.assembler.convertAsyncToDTO(this.queryService.deleteOne(id, this.convertFilterable(opts)))
+  public async deleteOne(id: number | string, opts?: DeleteOneOptions<DTO>): Promise<DTO> {
+    return this.assembler.convertToDTO(await this.queryService.deleteOne(id, this.convertFilterable(opts)))
   }
 
   public async findById(id: string | number, opts?: FindByIdOptions<DTO>): Promise<DTO | undefined> {
@@ -67,12 +67,12 @@ export class AssemblerQueryService<DTO, Entity, C = DeepPartial<DTO>, CE = DeepP
     return this.assembler.convertToDTO(entity)
   }
 
-  public getById(id: string | number, opts?: GetByIdOptions<DTO>): Promise<DTO> {
-    return this.assembler.convertAsyncToDTO(this.queryService.getById(id, this.convertFilterable(opts)))
+  public async getById(id: string | number, opts?: GetByIdOptions<DTO>): Promise<DTO> {
+    return this.assembler.convertToDTO(await this.queryService.getById(id, this.convertFilterable(opts)))
   }
 
-  public query(query: Query<DTO>, opts?: QueryOptions<DTO>): Promise<DTO[]> {
-    return this.assembler.convertAsyncToDTOs(this.queryService.query(this.assembler.convertQuery(query), opts as never))
+  public async query(query: Query<DTO>, opts?: QueryOptions<DTO>): Promise<DTO[]> {
+    return this.assembler.convertToDTOs(await this.queryService.query(this.assembler.convertQuery(query), opts as never))
   }
 
   public async aggregate(
@@ -128,7 +128,7 @@ export class AssemblerQueryService<DTO, Entity, C = DeepPartial<DTO>, CE = DeepP
     query: Query<Relation>
   ): Promise<Relation[] | Map<DTO, Relation[]>> {
     if (Array.isArray(dto)) {
-      const entities = this.assembler.convertToEntities(dto)
+      const entities = await this.assembler.convertToEntities(dto)
       const relationMap = await this.queryService.queryRelations(RelationClass, relationName, entities, query)
 
       return entities.reduce((map, e, index) => {
@@ -140,7 +140,7 @@ export class AssemblerQueryService<DTO, Entity, C = DeepPartial<DTO>, CE = DeepP
       }, new Map<DTO, Relation[]>())
     }
 
-    return this.queryService.queryRelations(RelationClass, relationName, this.assembler.convertToEntity(dto), query)
+    return this.queryService.queryRelations(RelationClass, relationName, await this.assembler.convertToEntity(dto), query)
   }
 
   public countRelations<Relation>(
@@ -164,7 +164,7 @@ export class AssemblerQueryService<DTO, Entity, C = DeepPartial<DTO>, CE = DeepP
     filter: Filter<Relation>
   ): Promise<number | Map<DTO, number>> {
     if (Array.isArray(dto)) {
-      const entities = this.assembler.convertToEntities(dto)
+      const entities = await this.assembler.convertToEntities(dto)
       const relationMap = await this.queryService.countRelations(RelationClass, relationName, entities, filter)
       return entities.reduce((map, e, index) => {
         const entry = relationMap.get(e) ?? 0
@@ -172,7 +172,7 @@ export class AssemblerQueryService<DTO, Entity, C = DeepPartial<DTO>, CE = DeepP
         return map
       }, new Map<DTO, number>())
     }
-    return this.queryService.countRelations(RelationClass, relationName, this.assembler.convertToEntity(dto), filter)
+    return this.queryService.countRelations(RelationClass, relationName, await this.assembler.convertToEntity(dto), filter)
   }
 
   /**
@@ -210,71 +210,71 @@ export class AssemblerQueryService<DTO, Entity, C = DeepPartial<DTO>, CE = DeepP
     opts?: FindRelationOptions<Relation>
   ): Promise<(Relation | undefined) | Map<DTO, Relation | undefined>> {
     if (Array.isArray(dto)) {
-      const entities = this.assembler.convertToEntities(dto)
+      const entities = await this.assembler.convertToEntities(dto)
       const relationMap = await this.queryService.findRelation(RelationClass, relationName, entities, opts)
       return entities.reduce((map, e, index) => {
         map.set(dto[index], relationMap.get(e))
         return map
       }, new Map<DTO, Relation | undefined>())
     }
-    return this.queryService.findRelation(RelationClass, relationName, this.assembler.convertToEntity(dto))
+    return this.queryService.findRelation(RelationClass, relationName, await this.assembler.convertToEntity(dto))
   }
 
-  public removeRelation<Relation>(
+  public async removeRelation<Relation>(
     relationName: string,
     id: string | number,
     relationId: string | number,
     opts?: ModifyRelationOptions<DTO, Relation>
   ): Promise<DTO> {
-    return this.assembler.convertAsyncToDTO(
-      this.queryService.removeRelation(relationName, id, relationId, this.convertModifyRelationsOptions(opts))
+    return this.assembler.convertToDTO(
+      await this.queryService.removeRelation(relationName, id, relationId, this.convertModifyRelationsOptions(opts))
     )
   }
 
-  public removeRelations<Relation>(
+  public async removeRelations<Relation>(
     relationName: string,
     id: string | number,
     relationIds: (string | number)[],
     opts?: ModifyRelationOptions<DTO, Relation>
   ): Promise<DTO> {
-    return this.assembler.convertAsyncToDTO(
-      this.queryService.removeRelations(relationName, id, relationIds, this.convertModifyRelationsOptions(opts))
+    return this.assembler.convertToDTO(
+      await this.queryService.removeRelations(relationName, id, relationIds, this.convertModifyRelationsOptions(opts))
     )
   }
 
-  public setRelations<Relation>(
+  public async setRelations<Relation>(
     relationName: string,
     id: string | number,
     relationIds: (string | number)[],
     opts?: ModifyRelationOptions<DTO, Relation>
   ): Promise<DTO> {
-    return this.assembler.convertAsyncToDTO(
-      this.queryService.setRelations(relationName, id, relationIds, this.convertModifyRelationsOptions(opts))
+    return this.assembler.convertToDTO(
+      await this.queryService.setRelations(relationName, id, relationIds, this.convertModifyRelationsOptions(opts))
     )
   }
 
-  public setRelation<Relation>(
+  public async setRelation<Relation>(
     relationName: string,
     id: string | number,
     relationId: string | number,
     opts?: ModifyRelationOptions<DTO, Relation>
   ): Promise<DTO> {
-    return this.assembler.convertAsyncToDTO(
-      this.queryService.setRelation(relationName, id, relationId, this.convertModifyRelationsOptions(opts))
+    return this.assembler.convertToDTO(
+      await this.queryService.setRelation(relationName, id, relationId, this.convertModifyRelationsOptions(opts))
     )
   }
 
-  public updateMany(update: U, filter: Filter<DTO>): Promise<UpdateManyResponse> {
+  public async updateMany(update: U, filter: Filter<DTO>): Promise<UpdateManyResponse> {
     return this.queryService.updateMany(
-      this.assembler.convertToUpdateEntity(update),
+      await this.assembler.convertToUpdateEntity(update),
       // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
       this.assembler.convertQuery({ filter }).filter
     )
   }
 
-  public updateOne(id: string | number, update: U, opts?: UpdateOneOptions<DTO>): Promise<DTO> {
-    return this.assembler.convertAsyncToDTO(
-      this.queryService.updateOne(id, this.assembler.convertToUpdateEntity(update), this.convertFilterable(opts))
+  public async updateOne(id: string | number, update: U, opts?: UpdateOneOptions<DTO>): Promise<DTO> {
+    return this.assembler.convertToDTO(
+      await this.queryService.updateOne(id, await this.assembler.convertToUpdateEntity(update), this.convertFilterable(opts))
     )
   }
 
@@ -300,7 +300,7 @@ export class AssemblerQueryService<DTO, Entity, C = DeepPartial<DTO>, CE = DeepP
     aggregate: AggregateQuery<Relation>
   ): Promise<AggregateResponse<Relation>[] | Map<DTO, AggregateResponse<Relation>[]>> {
     if (Array.isArray(dto)) {
-      const entities = this.assembler.convertToEntities(dto)
+      const entities = await this.assembler.convertToEntities(dto)
       const relationMap = await this.queryService.aggregateRelations(RelationClass, relationName, entities, filter, aggregate)
       return entities.reduce((map, e, index) => {
         const entry = relationMap.get(e) ?? []
@@ -311,7 +311,7 @@ export class AssemblerQueryService<DTO, Entity, C = DeepPartial<DTO>, CE = DeepP
     return this.queryService.aggregateRelations(
       RelationClass,
       relationName,
-      this.assembler.convertToEntity(dto),
+      await this.assembler.convertToEntity(dto),
       filter,
       aggregate
     )